+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/ervitis/filecache
+
+go 1.18
